requirements.txt
# Core dependencies
requests>=2.28.0
click>=8.0.0
tqdm>=4.64.0
toml>=0.10.2
python-dotenv>=1.0.0
GitPython>=3.1.0
